+++ packages/pdfkit/src/mixins/outline.js
@@ -9,7 +9,9 @@ export default {
     this.outline.endOutline();
     if (this.outline.children.length > 0) {
       this._root.data.Outlines = this.outline.dictionary;
-      return (this._root.data.PageMode = 'UseOutlines');
+      /* Custom fork start */
+      return (this._root.data.PageMode = this._root.data.PageMode || 'UseOutlines');
+      /* Custom fork end */
     }
   }
 };