+++ packages/renderer/index.d.ts
@@ -611,6 +611,7 @@ declare namespace ReactPDF {
         event: React.MouseEvent<HTMLAnchorElement, MouseEvent>,
         instance: UsePDFInstance,
       ) => void);
+    ref?: React.Ref<HTMLAnchorElement> | undefined;
   }
 
   /**